Move Table class to separate file
