+++ scraper.rb
@@ -4,27 +4,9 @@
 require 'scraperwiki'
 require 'nokogiri'
 require 'open-uri/cached'
+require_relative 'lib/table'
 
 require 'pry'
-
-class Table
-  def initialize(node)
-    @table = node
-  end
-
-  def rows
-    constituency = nil
-    table.xpath('.//tr[td]').map do |tr|
-      tds = tr.xpath('./td')
-      constituency = tds.shift.text.strip.gsub("\n",' — ') if tds.first[:rowspan]
-      Row.new(tds).to_h.merge(constituency: constituency)
-    end
-  end
-
-  private
-
-  attr_reader :table
-end
 
 class Row
   def initialize(tds)
